+++ src/modules/auth/components/AuthTextField/AuthTextField.tsx
@@ -35,7 +35,7 @@ export const AuthTextField = ({
       name={name}
       className={classNames(
         className,
-        'pw-rounded-[7px] !pw-text-[15px] !pw-font-normal !pw-leading-[18px] pw-mt-1 !pw-outline-1',
+        'pw-rounded-[7px] !pw-font-normal !pw-leading-[18px] pw-mt-1 !pw-outline-1',
 
         getOutlineClassName()
       )}
@@ -44,7 +44,7 @@ export const AuthTextField = ({
       autoComplete={autoComplete}
       classes={{
         rootDisabled: '!pw-outline-[#676767] pw-bg-transparent',
-        input: `!pw-px-[10px] !pw-py-[14px] placeholder:!pw-text-[#777E8F] !pw-text-[13px] !pw-leading-[18px] autofill:pw-bg-transparent disabled:pw-text-[#353945] disabled:pw-bg-[#EFEFEF] pw-text-fill-[#353945] autofill:pw-bg-transparent autofill:pw-shadow-[0_0_0_30px_#ffffff_inset] ${
+        input: `!pw-px-[10px] !pw-py-[14px] placeholder:!pw-text-[#777E8F] !pw-leading-[18px] autofill:pw-bg-transparent disabled:pw-text-[#353945] disabled:pw-bg-[#EFEFEF] pw-text-fill-[#353945] autofill:pw-bg-transparent autofill:pw-shadow-[0_0_0_30px_#ffffff_inset] ${
           invalid ? '!pw-text-[#C63535]' : '!pw-text-[#353945]'
         }`,
       }}